move nodes back to their original location
