+++ src/plugins/dataflow/nodes/base-node.ts
@@ -1,11 +1,11 @@
 import { Instance, types } from "mobx-state-tree";
-import { defaultMinigraphOptions } from "../../nodes/dataflow-node-plot";
-import { kMaxNodeValues } from "../../model/utilities/node";
+import { defaultMinigraphOptions } from "./dataflow-node-plot";
+import { kMaxNodeValues } from "../model/utilities/node";
 import { ClassicPreset } from "rete";
 import { Socket } from "rete/_types/presets/classic";
-import { DataflowProgramChange } from "../../dataflow-logger";
-import { INodeServices } from "../service-types";
-import { Schemes } from "../rete-scheme";
+import { DataflowProgramChange } from "../dataflow-logger";
+import { INodeServices } from "../rete/service-types";
+import { Schemes } from "../rete/rete-scheme";
 
 export type NoInputs = Record<string, never>;
 export type NoOutputs = Record<string, never>;
