+++ src/plugins/dataflow/nodes/controls/demo-output-control.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { advancedGrabberFrames, fanFrames, fanHousing, fanMotor, grabberCordFrames, grabberFrames,
   grabberPaddle, humidifier, lightBulbOff, lightBulbOn } from "./demo-output-control-assets";
 import { HumidifierAnimation } from "./demo-outputs/humidifier-animation";
-import { IDemoOutputNodeModel } from "../../rete/nodes/demo-output-node";
+import { IDemoOutputNodeModel } from "../demo-output-node";
 
 import "./demo-output-control.scss";
