+++ src/plugins/dataflow/nodes/controls/input-value-control.tsx
@@ -3,7 +3,7 @@ import "./value-control.sass";
 import React from "react";
 import { observer } from "mobx-react";
 import { action, computed, makeObservable, observable } from "mobx";
-import { IBaseNode, IBaseNodeModel } from "../../rete/nodes/base-node";
+import { IBaseNode, IBaseNodeModel } from "../base-node";
 import { MinigraphOptions, defaultMinigraphOptions } from "../dataflow-node-plot";
 import { PlotButtonControl, PlotButtonControlComponent } from "./plot-button-control";
