+++ src/plugins/dataflow/nodes/dataflow-node-plot.tsx
@@ -4,7 +4,7 @@ import { ChartOptions, ChartData, ChartDataSets } from "chart.js";
 import { kMaxNodeValues, NodePlotColor } from "../model/utilities/node";
 
 import "./dataflow-node.scss";
-import { IBaseNodeModel } from "../rete/nodes/base-node";
+import { IBaseNodeModel } from "./base-node";
 import { observer } from "mobx-react";
 import { useStopEventPropagation } from "./controls/custom-hooks";
