+++ src/plugins/dataflow/nodes/generator-node.ts
@@ -1,15 +1,15 @@
 import { ClassicPreset } from "rete";
 import { Instance } from "mobx-state-tree";
-import { numSocket } from "../num-socket";
-import { INumberUnitsControl, NumberUnitsControl } from "../../nodes/controls/num-units-control";
+import { numSocket } from "../rete/num-socket";
+import { INumberUnitsControl, NumberUnitsControl } from "./controls/num-units-control";
 import { BaseNode, BaseNodeModel, NoInputs } from "./base-node";
-import { NodeGeneratorTypes, NodePeriodUnits } from "../../model/utilities/node";
-import { DropdownListControl, IDropdownListControl } from "../../nodes/controls/dropdown-list-control";
-import { ValueControl } from "../../nodes/controls/value-control";
-import { PlotButtonControl } from "../../nodes/controls/plot-button-control";
-import { typeField } from "../../../../utilities/mst-utils";
-import { INodeServices } from "../service-types";
-import { INumberControl, NumberControl } from "../../nodes/controls/num-control";
+import { NodeGeneratorTypes, NodePeriodUnits } from "../model/utilities/node";
+import { DropdownListControl, IDropdownListControl } from "./controls/dropdown-list-control";
+import { ValueControl } from "./controls/value-control";
+import { PlotButtonControl } from "./controls/plot-button-control";
+import { typeField } from "../../../utilities/mst-utils";
+import { INodeServices } from "../rete/service-types";
+import { INumberControl, NumberControl } from "./controls/num-control";
 
 export const GeneratorNodeModel = BaseNodeModel.named("GeneratorNodeModel")
 .props({
